fix: MatchStore class 버그 픽스
테스트 시 발견되었던 MatchStroe class 버그를 픽스
	1. getYield 메소드의 return 값이 long이였던 것 수정
	2. MatchStore에서 addMatchResult가 불리지 않았을 때 getYield 메소드의 리턴 값이 div 0 또는 NaN으로 출력되었던것 수정

+++ src/main/java/controller/ResultController.java
@@ -24,7 +24,7 @@ public class ResultController {
         OutputView.printPreResult();
     }
 
-    public static void printYield(long yield) {
+    public static void printYield(float yield) {
         OutputView.printYield(yield);
     }
