+++ src/main/java/view/OutputView.java
@@ -10,6 +10,7 @@ public class OutputView {
     public static void printPayInputError() {
         System.out.println("적어도 한개의 복권을 구매해주세요 (개당 1000원)");
     }
+
     public static void printManualNumberError(int totalNum) {
         System.out.printf("0부터 %d까지의 정수 중 하나의 수를 입력해주세요\n", totalNum);
     }
@@ -22,9 +23,13 @@ public class OutputView {
         System.out.println("1부터 45까지의 정수 중에서 당첨번호에 포함되지 않는 수 하나를 입력해주세요.");
     }
 
-    public static void announceBeforeManualLine() {System.out.println("수동으로 구매할 번호를 입력해 주세요.");}
+    public static void announceBeforeManualLine() {
+        System.out.println("수동으로 구매할 번호를 입력해 주세요.");
+    }
 
-    public static void announceBeforeWinningLine() {System.out.println("지난 주 당첨 번호를 입력해 주세요.");}
+    public static void announceBeforeWinningLine() {
+        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
+    }
 
     public static void printLottoLine(String lottoLineStr) {
         System.out.println(lottoLineStr);
@@ -49,7 +54,7 @@ public class OutputView {
         System.out.printf("%d개 일치 (%d원)- %d개\n", result.getMatchNum(), result.getPrice(), result.getNumLotto());
     }
 
-    public static void printYield(Long yield) {
+    public static void printYield(float yield) {
         System.out.printf("총 수익률은 %d%%입니다.", yield);
     }
 }
